feat(sstd): change semantics of state pattern matcher
Previously patterns in state pattern matcher were
completely independent of eachother
(if the state is not counted).

Now it matches only one pattern at the same time,
i.e. new pattern is the first pattern which predicate is true and
it is matched after end of previous pattern is called.